+++ src/__mocks__/server.ts
@@ -14,8 +14,10 @@ import {
   pemAccountGuarded,
   faucetSettings,
   keystoreWalletCollections,
-  pendingTransactionKeystoreWallet
+  pendingTransactionKeystoreWallet,
+  emptyWalletAccount
 } from './data';
+import { issueContract } from './data/issueContract';
 import { mockResponse } from './serverUtils';
 
 export const handlers = [
@@ -27,6 +29,10 @@ export const handlers = [
   http.get(
     `${testNetwork.apiAddress}/accounts/${keystoreAccount.address}`,
     mockResponse(keystoreAccount)
+  ),
+  http.get(
+    `${testNetwork.apiAddress}/accounts/${emptyWalletAccount.address}`,
+    mockResponse(emptyWalletAccount)
   ),
   http.get(
     `${testNetwork.apiAddress}/accounts/${keystoreAccount.address}/tokens`,
@@ -67,6 +73,10 @@ export const handlers = [
   http.get(
     `${testNetwork.apiAddress}/accounts/${pemAccountGuarded.address}/nfts`,
     mockResponse(pemWalletNfts)
+  ),
+  http.get(
+    `${testNetwork.apiAddress}/accounts/${issueContract.address}`,
+    mockResponse(issueContract)
   ),
   http.get(
     `${testNetwork.extrasApiAddress}/faucet/settings`,