+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- [Fixed sovereign contract address in register token](https://github.com/multiversx/mx-lite-wallet-dapp/pull/72)
+- [Fixed sovereign contract address in register token](https://github.com/multiversx/mx-lite-wallet-dapp/pull/73)
 
 ## [[1.0.4](https://github.com/multiversx/mx-lite-wallet-dapp/pull/71)] - 2025-01-30